Uppercase "Assignment"
Ticket, Backend, MMLogic is all uppercased - becasue they are special
nouns and concept in this system. Assignment is the same, as we should
upper case it when we mean a Open Match Assignment, rather than a
colloquial assignment of a different variety.

I personally found it tricky to remember it was a core concept
otherwise.